Crypto.HMAC

Description

Documentation

hmac :: Hash c d => MacKey c d -> ByteString -> d

Message authentication code calculation for lazy bytestrings. hmac k msg will compute an authentication code for msg using key k

hmac' :: Hash c d => MacKey c d -> ByteString -> d

hmac k msg will compute an authentication code for msg using key k

newtype MacKey c d

A key carrying phantom types c and d, forcing the key data to only be used by particular hash algorithms.

Constructors

MacKey ByteString