Crypto.Classes.Exceptions

Description

The module mirrors Crypto.Classes except that errors are thrown as exceptions instead of having returning types of Either error result or Maybe result.

NB This module is experimental and might go away or be re-arranged.

Documentation

class (Serialize d, Eq d, Ord d) => Hash ctx d | d -> ctx, ctx -> d where

The Hash class is intended as the generic interface targeted by maintainers of Haskell digest implementations. Using this generic interface, higher level functions such as hash and hash' provide a useful API for comsumers of hash implementations.

Any instantiated implementation must handle unaligned data.

Minimum complete definition: outputLength, blockLength, initialCtx, updateCtx, and finalize.

Minimal complete definition

outputLength, blockLength, initialCtx, updateCtx, finalize

Methods

outputLength :: Tagged d BitLength

blockLength :: Tagged d BitLength

initialCtx :: ctx

updateCtx :: ctx -> ByteString -> ctx

finalize :: ctx -> ByteString -> d

hash :: Hash ctx d => ByteString -> d

Hash a lazy ByteString, creating a digest

hash' :: Hash ctx d => ByteString -> d

Hash a strict ByteString, creating a digest

hashFunc' :: Hash c d => d -> ByteString -> d

Obtain a strict hash function whose result is the same type as the given digest, which is discarded. If the type is already inferred then consider using the hash' function instead.

hashFunc :: Hash c d => d -> ByteString -> d

Obtain a lazy hash function whose result is the same type as the given digest, which is discarded. If the type is already inferred then consider using the hash function instead.

class Serialize k => BlockCipher k where

The BlockCipher class is intended as the generic interface targeted by maintainers of Haskell cipher implementations.

Minimum complete definition: blockSize, encryptBlock, decryptBlock, buildKey, and keyLength.

Instances must handle unaligned data

Minimal complete definition

blockSize, encryptBlock, decryptBlock, buildKey, keyLength

Methods

blockSize :: Tagged k BitLength

encryptBlock :: k -> ByteString -> ByteString

decryptBlock :: k -> ByteString -> ByteString

keyLength :: Tagged k BitLength

ecb :: k -> ByteString -> ByteString

Electronic Cookbook (encryption)

unEcb :: k -> ByteString -> ByteString

Electronic Cookbook (decryption)

cbc :: k -> IV k -> ByteString -> (ByteString, IV k)

Cipherblock Chaining (encryption)

unCbc :: k -> IV k -> ByteString -> (ByteString, IV k)

Cipherblock Chaining (decryption)

ctr :: k -> IV k -> ByteString -> (ByteString, IV k)

Counter (encryption)

unCtr :: k -> IV k -> ByteString -> (ByteString, IV k)

Counter (decryption)

ctrLazy :: k -> IV k -> ByteString -> (ByteString, IV k)

Counter (encryption)

unCtrLazy :: k -> IV k -> ByteString -> (ByteString, IV k)

Counter (decryption)

cfb :: k -> IV k -> ByteString -> (ByteString, IV k)

Ciphertext feedback (encryption)

unCfb :: k -> IV k -> ByteString -> (ByteString, IV k)

Ciphertext feedback (decryption)

ofb :: k -> IV k -> ByteString -> (ByteString, IV k)

Output feedback (encryption)

unOfb :: k -> IV k -> ByteString -> (ByteString, IV k)

Output feedback (decryption)

cbcLazy :: k -> IV k -> ByteString -> (ByteString, IV k)

Cipher block chaining encryption for lazy bytestrings

unCbcLazy :: k -> IV k -> ByteString -> (ByteString, IV k)

Cipher block chaining decryption for lazy bytestrings

sivLazy :: k -> k -> [ByteString] -> ByteString -> Maybe ByteString

SIV (Synthetic IV) mode for lazy bytestrings. The third argument is the optional list of bytestrings to be authenticated but not encrypted As required by the specification this algorithm may return nothing when certain constraints aren't met.

unSivLazy :: k -> k -> [ByteString] -> ByteString -> Maybe ByteString

SIV (Synthetic IV) for lazy bytestrings. The third argument is the optional list of bytestrings to be authenticated but not encrypted. As required by the specification this algorithm may return nothing when authentication fails.

siv :: k -> k -> [ByteString] -> ByteString -> Maybe ByteString

SIV (Synthetic IV) mode for strict bytestrings. First argument is the optional list of bytestrings to be authenticated but not encrypted. As required by the specification this algorithm may return nothing when certain constraints aren't met.

unSiv :: k -> k -> [ByteString] -> ByteString -> Maybe ByteString

SIV (Synthetic IV) for strict bytestrings First argument is the optional list of bytestrings to be authenticated but not encrypted As required by the specification this algorithm may return nothing when authentication fails.

ecbLazy :: k -> ByteString -> ByteString

Cook book mode - not really a mode at all. If you don't know what you're doing, don't use this mode^H^H^H^H library.

unEcbLazy :: k -> ByteString -> ByteString

ECB decrypt, complementary to ecb.

cfbLazy :: k -> IV k -> ByteString -> (ByteString, IV k)

Ciphertext feed-back encryption mode for lazy bytestrings (with s == blockSize)

unCfbLazy :: k -> IV k -> ByteString -> (ByteString, IV k)

Ciphertext feed-back decryption mode for lazy bytestrings (with s == blockSize)

ofbLazy :: k -> IV k -> ByteString -> (ByteString, IV k)

Output feedback mode for lazy bytestrings

unOfbLazy :: k -> IV k -> ByteString -> (ByteString, IV k)

Output feedback mode for lazy bytestrings

blockSize :: BlockCipher k => Tagged k BitLength

encryptBlock :: BlockCipher k => k -> ByteString -> ByteString

decryptBlock :: BlockCipher k => k -> ByteString -> ByteString

keyLength :: BlockCipher k => Tagged k BitLength

getIVIO :: BlockCipher k => IO (IV k)

Obtain an IV using the system entropy (see Entropy)

blockSizeBytes :: BlockCipher k => Tagged k ByteLength

The number of bytes in a block cipher block

keyLengthBytes :: BlockCipher k => Tagged k ByteLength

The number of bytes in a block cipher key (assuming it is an even multiple of 8 bits)

buildKeyIO :: BlockCipher k => IO k

Build a symmetric key using the system entropy (see Entropy)

class AsymCipher p v | p -> v, v -> p where

Asymetric ciphers (common ones being RSA or EC based)

Minimal complete definition

buildKeyPair, encryptAsym, decryptAsym, publicKeyLength, privateKeyLength

Methods

publicKeyLength :: p -> BitLength

privateKeyLength :: v -> BitLength

publicKeyLength :: AsymCipher p v => p -> BitLength

privateKeyLength :: AsymCipher p v => v -> BitLength

buildKeyPairIO :: AsymCipher p v => BitLength -> IO (Either GenError (p, v))

Build a pair of asymmetric keys using the system random generator.

class (Serialize p, Serialize v) => Signing p v | p -> v, v -> p where

A class for signing operations which inherently can not be as generic as asymetric ciphers (ex: DSA).

Minimal complete definition

sign, verify, buildSigningPair, signingKeyLength, verifyingKeyLength

Methods

verify :: p -> ByteString -> ByteString -> Bool

signingKeyLength :: v -> BitLength

verifyingKeyLength :: p -> BitLength

signingKeyLength :: Signing p v => v -> BitLength

verifyingKeyLength :: Signing p v => p -> BitLength

verify :: Signing p v => p -> ByteString -> ByteString -> Bool

incIV :: BlockCipher k => IV k -> IV k

Increase an IV by one. This is way faster than decoding, increasing, encoding

zeroIV :: BlockCipher k => IV k

Obtain an IV made only of zeroes

class CryptoRandomGen g where

A class of random bit generators that allows for the possibility of failure, reseeding, providing entropy at the same time as requesting bytes

Minimum complete definition: newGen, genSeedLength, genBytes, reseed, reseedInfo, reseedPeriod.

Minimal complete definition

newGen, genSeedLength, genBytes, reseedInfo, reseedPeriod, reseed

Methods

genSeedLength :: Tagged g ByteLength

Length of input entropy necessary to instantiate or reseed a generator

reseedInfo :: g -> ReseedInfo

Indicates how soon a reseed is needed

reseedPeriod :: g -> ReseedInfo

Indicates the period between reseeds (constant for most generators).

newGenIO :: IO g

By default this uses System.Entropy to obtain entropy for newGen.

genSeedLength :: CryptoRandomGen g => Tagged g ByteLength

Length of input entropy necessary to instantiate or reseed a generator

reseedInfo :: CryptoRandomGen g => g -> ReseedInfo

Indicates how soon a reseed is needed

reseedPeriod :: CryptoRandomGen g => g -> ReseedInfo

Indicates the period between reseeds (constant for most generators).

newGenIO :: CryptoRandomGen g => IO g

By default this uses System.Entropy to obtain entropy for newGen.

data GenError

Generator failures should always return the appropriate GenError. Note GenError in an instance of exception but wether or not an exception is thrown depends on if the selected generator (read: if you don't want execptions from code that uses throw then pass in a generator that never has an error for the used functions)

Constructors

GenErrorOther String
Misc

RequestedTooManyBytes
Requested more bytes than a single pass can generate (The maximum request is generator dependent)

RangeInvalid
When using genInteger g (l,h) and logBase 2 (h - l) > (maxBound :: Int).

NeedReseed
Some generators cease operation after too high a count without a reseed (ex: NIST SP 800-90)

NotEnoughEntropy
For instantiating new generators (or reseeding)

NeedsInfiniteSeed
This generator can not be instantiated or reseeded with a finite seed (ex: SystemRandom)

data ReseedInfo

Constructors

InXBytes !Word64
Generator needs reseeded in X bytes

InXCalls !Word64
Generator needs reseeded in X calls

NotSoon
The bound is over 2^64 bytes or calls

Never
This generator never reseeds (ex: SystemRandom)

data CipherError

Constructors

GenError GenError

KeyGenFailure

ecb :: BlockCipher k => k -> ByteString -> ByteString

Electronic Cookbook (encryption)

unEcb :: BlockCipher k => k -> ByteString -> ByteString

Electronic Cookbook (decryption)

cbc :: BlockCipher k => k -> IV k -> ByteString -> (ByteString, IV k)

Cipherblock Chaining (encryption)

unCbc :: BlockCipher k => k -> IV k -> ByteString -> (ByteString, IV k)

Cipherblock Chaining (decryption)

ctr :: BlockCipher k => k -> IV k -> ByteString -> (ByteString, IV k)

Counter (encryption)

unCtr :: BlockCipher k => k -> IV k -> ByteString -> (ByteString, IV k)

Counter (decryption)

ctrLazy :: BlockCipher k => k -> IV k -> ByteString -> (ByteString, IV k)

Counter (encryption)

unCtrLazy :: BlockCipher k => k -> IV k -> ByteString -> (ByteString, IV k)

Counter (decryption)

cfb :: BlockCipher k => k -> IV k -> ByteString -> (ByteString, IV k)

Ciphertext feedback (encryption)

unCfb :: BlockCipher k => k -> IV k -> ByteString -> (ByteString, IV k)

Ciphertext feedback (decryption)

ofb :: BlockCipher k => k -> IV k -> ByteString -> (ByteString, IV k)

Output feedback (encryption)

unOfb :: BlockCipher k => k -> IV k -> ByteString -> (ByteString, IV k)

Output feedback (decryption)

cbcLazy :: BlockCipher k => k -> IV k -> ByteString -> (ByteString, IV k)

Cipher block chaining encryption for lazy bytestrings

unCbcLazy :: BlockCipher k => k -> IV k -> ByteString -> (ByteString, IV k)

Cipher block chaining decryption for lazy bytestrings

sivLazy :: BlockCipher k => k -> k -> [ByteString] -> ByteString -> Maybe ByteString

SIV (Synthetic IV) mode for lazy bytestrings. The third argument is the optional list of bytestrings to be authenticated but not encrypted As required by the specification this algorithm may return nothing when certain constraints aren't met.

unSivLazy :: BlockCipher k => k -> k -> [ByteString] -> ByteString -> Maybe ByteString

SIV (Synthetic IV) for lazy bytestrings. The third argument is the optional list of bytestrings to be authenticated but not encrypted. As required by the specification this algorithm may return nothing when authentication fails.

siv :: BlockCipher k => k -> k -> [ByteString] -> ByteString -> Maybe ByteString

SIV (Synthetic IV) mode for strict bytestrings. First argument is the optional list of bytestrings to be authenticated but not encrypted. As required by the specification this algorithm may return nothing when certain constraints aren't met.

unSiv :: BlockCipher k => k -> k -> [ByteString] -> ByteString -> Maybe ByteString

SIV (Synthetic IV) for strict bytestrings First argument is the optional list of bytestrings to be authenticated but not encrypted As required by the specification this algorithm may return nothing when authentication fails.

ecbLazy :: BlockCipher k => k -> ByteString -> ByteString

Cook book mode - not really a mode at all. If you don't know what you're doing, don't use this mode^H^H^H^H library.

unEcbLazy :: BlockCipher k => k -> ByteString -> ByteString

ECB decrypt, complementary to ecb.

cfbLazy :: BlockCipher k => k -> IV k -> ByteString -> (ByteString, IV k)

Ciphertext feed-back encryption mode for lazy bytestrings (with s == blockSize)

unCfbLazy :: BlockCipher k => k -> IV k -> ByteString -> (ByteString, IV k)

Ciphertext feed-back decryption mode for lazy bytestrings (with s == blockSize)

ofbLazy :: BlockCipher k => k -> IV k -> ByteString -> (ByteString, IV k)

Output feedback mode for lazy bytestrings

unOfbLazy :: BlockCipher k => k -> IV k -> ByteString -> (ByteString, IV k)

Output feedback mode for lazy bytestrings

buildKey :: BlockCipher k => ByteString -> k

Key construction from raw material (typically including key expansion)

This is a wrapper that can throw a CipherError on exception.

getIV :: (BlockCipher k, CryptoRandomGen g) => g -> (IV k, g)

Random IV generation

This is a wrapper that can throw a GenError on exception.

buildKeyGen :: (CryptoRandomGen g, BlockCipher k) => g -> (k, g)

Symmetric key generation

This is a wrapper that can throw a GenError on exception.

buildKeyPair :: (CryptoRandomGen g, AsymCipher p v) => g -> BitLength -> ((p, v), g)

Asymetric key generation

This is a wrapper that can throw a GenError on exception.

encryptAsym :: (CryptoRandomGen g, AsymCipher p v) => g -> p -> ByteString -> (ByteString, g)

Asymmetric encryption

This is a wrapper that can throw a GenError on exception.

decryptAsym :: (CryptoRandomGen g, AsymCipher p v) => g -> v -> ByteString -> (ByteString, g)

Asymmetric decryption

This is a wrapper that can throw a GenError on exception.

newGen :: CryptoRandomGen g => ByteString -> g

Instantiate a new random bit generator. The provided bytestring should be of length >= genSeedLength. If the bytestring is shorter then the call may fail (suggested error: NotEnoughEntropy). If the bytestring is of sufficent length the call should always succeed.

This is a wrapper that can throw GenError types as exceptions.

genBytes :: CryptoRandomGen g => ByteLength -> g -> (ByteString, g)

genBytes len g generates a random ByteString of length len and new generator. The MonadCryptoRandom package has routines useful for converting the ByteString to commonly needed values (but cereal or other deserialization libraries would also work).

This is a wrapper that can throw GenError types as exceptions.

genBytesWithEntropy :: CryptoRandomGen g => ByteLength -> ByteString -> g -> (ByteString, g)

genBytesWithEntropy g i entropy generates i random bytes and use the additional input entropy in the generation of the requested data to increase the confidence our generated data is a secure random stream.

This is a wrapper that can throw GenError types as exceptions.

reseed :: CryptoRandomGen g => ByteString -> g -> g

If the generator has produced too many random bytes on its existing seed it will throw a NeedReseed exception. In that case, reseed the generator using this function and a new high-entropy seed of length >= genSeedLength. Using bytestrings that are too short can result in an exception (NotEnoughEntropy).

splitGen :: CryptoRandomGen g => g -> (g, g)

While the safety and wisdom of a splitting function depends on the properties of the generator being split, several arguments from informed people indicate such a function is safe for NIST SP 800-90 generators. (see libraries@haskell.org discussion around Sept, Oct 2010). You can find implementations of such generators in the DRBG package.

This is a wrapper for splitGen which throws errors as exceptions.